+++ webapp/controller/AvioesController.php
@@ -10,22 +10,33 @@ class AvioesController extends BaseController implements ResourceControllerInter
 
     public function index()
     {
-        $avioes = Avioes::all();
-        return View::make('avioes.index', ['avioes' => $avioes]);
+        if(isset($_SESSION['username'])) {
+            $avioes = Avioes::all();
+            return View::make('avioes.index', ['avioes' => $avioes]);
+        }else{
+            Redirect::toRoute('utilizadores/index');
+        }
     }
 
     public function create()
     {
-        return View::make('avioes.create');
+        if(isset($_SESSION['username'])) {
+            return View::make('avioes.create');
+        }else{
+            Redirect::toRoute('utilizadores/index');
+        }
     }
 
     public function store()
     {
-        $aviao = Post::getAll();
-        $avioes = new Avioes($aviao);
-        $avioes->save();
-        Redirect::toRoute('avioes/index');
-
+        if(isset($_SESSION['username'])) {
+            $aviao = Post::getAll();
+            $avioes = new Avioes($aviao);
+            $avioes->save();
+            Redirect::toRoute('avioes/index');
+        }else{
+            Redirect::toRoute('utilizadores/index');
+        }
     }
 
     public function show($id)
@@ -35,23 +46,35 @@ class AvioesController extends BaseController implements ResourceControllerInter
 
     public function edit($id)
     {
-        $aviao = Avioes::find([$id]);
-        return View::make('avioes.edit', ['aviao' => [$aviao]]);
+        if(isset($_SESSION['username'])) {
+            $aviao = Avioes::find([$id]);
+            return View::make('avioes.edit', ['aviao' => [$aviao]]);
+        }else{
+            Redirect::toRoute('utilizadores/index');
+        }
     }
 
     public function update($id)
     {
-        $aviao = Avioes::find([$id]);
-        $aviao->update_attributes(Post::getAll());
-        $aviao->save();
-        Redirect::toRoute('avioes/index');
+        if(isset($_SESSION['username'])) {
+            $aviao = Avioes::find([$id]);
+            $aviao->update_attributes(Post::getAll());
+            $aviao->save();
+            Redirect::toRoute('avioes/index');
+        }else{
+            Redirect::toRoute('utilizadores/index');
+        }
     }
 
     public function destroy($id)
     {
-        $aviao = Avioes::find([$id]);
-        $aviao->delete();
-        Redirect::toRoute('avioes/index');
+        if(isset($_SESSION['username'])) {
+            $aviao = Avioes::find([$id]);
+            $aviao->delete();
+            Redirect::toRoute('avioes/index');
+        }else{
+            Redirect::toRoute('utilizadores/index');
+        }
     }
 }
 ?>
